+++ resources/lang/en/app.php
@@ -1,5 +1,7 @@
 <?php
 
 return [
+    'error' => 'Error',
     'name' => 'XNova',
+    'success' => 'Success',
 ];