+++ resources/lang/en/emails.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'password_reset' => [
+        'para1' => 'Hi :name!',
+        'para2' => 'You asked for the new password in :servername so here it is:',
+        'subject' => '[:servername] Password reset'
+    ],
+];