+++ resources/views/emails/password_reset.blade.php
@@ -1,0 +1,3 @@
+{{ trans('emails.password_reset.para1', ['name' => $user['username']]) }}<br><br>
+
+{{ trans('emails.password_reset.para2') }} {{ $password }}